Moved workers to separate package- I am planning on a sheetsdaemon and possibly other integrations. This looks much better.
