+++ src/main/kotlin/com/lowbudgetlcs/Application.kt
@@ -1,5 +1,7 @@
 package com.lowbudgetlcs
 
+import com.lowbudgetlcs.workers.StatDaemon
+import com.lowbudgetlcs.workers.TournamentEngine
 import io.ktor.server.application.*
 import io.ktor.util.logging.*
 import kotlinx.coroutines.launch
